feat(extension): make adaptive sync the default implementation
- Removed feature flag ENABLE_ADAPTIVE_SYNC
- Deleted old sync-manager implementation
- Renamed AdaptiveSyncManager to SyncManager
- Removed factory function createSyncManager
- Updated all imports and references
- Removed coverage exclusion for adaptive sync
- Updated documentation to reflect the change

The adaptive sync with intelligent batching, priority queues, and
operation deduplication is now the default and only sync implementation.
