+++ extension/src/sync/sync-manager.ts
@@ -15,53 +15,102 @@ interface SyncManagerConfig {
   browser: IBrowser;
 }
 
+enum OperationPriority {
+  CRITICAL = 0,
+  HIGH = 1,
+  NORMAL = 2,
+  LOW = 3,
+}
+
+interface QueuedOperation {
+  operation: CrdtOperation;
+  priority: OperationPriority;
+  timestamp: number;
+  dedupKey: string;
+}
+
+interface AdaptiveConfig {
+  activeIntervalMs: number;
+  idleIntervalMs: number;
+  errorBackoffMs: number;
+  maxBackoffMs: number;
+  activityThresholdMs: number;
+  batchDelays: Record<OperationPriority, number>;
+  maxQueueSize: number;
+  queueSizeThreshold: number;
+}
+
+const DEFAULT_ADAPTIVE_CONFIG: AdaptiveConfig = {
+  activeIntervalMs: 1000,
+  idleIntervalMs: 10000,
+  errorBackoffMs: 5000,
+  maxBackoffMs: 60000,
+  activityThresholdMs: 30000,
+  batchDelays: {
+    [OperationPriority.CRITICAL]: 50,
+    [OperationPriority.HIGH]: 200,
+    [OperationPriority.NORMAL]: 500,
+    [OperationPriority.LOW]: 1000,
+  },
+  maxQueueSize: 1000,
+  queueSizeThreshold: 50,
+};
+
 export class SyncManager {
-  private syncInterval: number;
+  private baseInterval: number;
+  private currentInterval: number;
   private syncTimer: number | null = null;
+  private batchTimer: number | null = null;
   private api: TanakaAPI;
   private windowTracker: WindowTracker;
   private browser: IBrowser;
   private deviceId: string;
   private lamportClock = 0n;
   private lastSyncClock: bigint | null = null;
-  private operationQueue: CrdtOperation[] = [];
+  private operationQueue: QueuedOperation[] = [];
   private isSyncing = false;
+  private lastActivityTime = 0;
+  private consecutiveErrors = 0;
+  private adaptiveConfig: AdaptiveConfig;
+  private pendingBatchPriority: OperationPriority | null = null;
 
   constructor(config: SyncManagerConfig) {
-    this.syncInterval = config.syncIntervalMs ?? 5000;
+    this.baseInterval = config.syncIntervalMs ?? 5000;
+    this.currentInterval = this.baseInterval;
     this.api = config.api;
     this.windowTracker = config.windowTracker;
     this.browser = config.browser;
     this.deviceId = config.deviceId ?? this.generateDeviceId();
+    this.adaptiveConfig = DEFAULT_ADAPTIVE_CONFIG;
   }
 
   private generateDeviceId(): string {
-    // Generate a unique device ID based on browser instance
     return `browser-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
   }
 
   start(): void {
     debugLog(
-      `Starting Sync Manager - syncInterval: ${this.syncInterval}, deviceId: ${this.deviceId}`,
+      `Starting Adaptive Sync Manager - baseInterval: ${this.baseInterval}, deviceId: ${this.deviceId}`,
     );
 
-    // Load persisted state
     this.loadPersistedState();
-
-    // Start sync timer
     this.scheduleSyncCheck();
   }
 
   stop(): void {
-    debugLog('Stopping Sync Manager');
+    debugLog('Stopping Adaptive Sync Manager');
     if (this.syncTimer) {
       clearTimeout(this.syncTimer);
       this.syncTimer = null;
     }
+    if (this.batchTimer) {
+      clearTimeout(this.batchTimer);
+      this.batchTimer = null;
+    }
   }
 
   async restart(): Promise<void> {
-    debugLog('Restarting Sync Manager');
+    debugLog('Restarting Adaptive Sync Manager');
     this.stop();
     this.start();
   }
@@ -112,16 +161,67 @@ export class SyncManager {
     }
   }
 
+  private calculateAdaptiveInterval(): number {
+    const now = Date.now();
+    const timeSinceActivity = now - this.lastActivityTime;
+    const isActive = timeSinceActivity < this.adaptiveConfig.activityThresholdMs;
+
+    let interval: number;
+
+    if (this.consecutiveErrors > 0) {
+      interval = Math.min(
+        this.adaptiveConfig.errorBackoffMs * Math.pow(2, this.consecutiveErrors - 1),
+        this.adaptiveConfig.maxBackoffMs,
+      );
+      debugLog(`Using error backoff interval: ${interval}ms (errors: ${this.consecutiveErrors})`);
+    } else if (isActive) {
+      interval = this.adaptiveConfig.activeIntervalMs;
+      debugLog(`Using active interval: ${interval}ms`);
+    } else {
+      interval = this.adaptiveConfig.idleIntervalMs;
+      debugLog(`Using idle interval: ${interval}ms`);
+    }
+
+    if (this.operationQueue.length > this.adaptiveConfig.queueSizeThreshold) {
+      interval = Math.min(interval, this.adaptiveConfig.activeIntervalMs);
+      debugLog(`Queue size ${this.operationQueue.length} exceeds threshold, using faster interval`);
+    }
+
+    return interval;
+  }
+
   private scheduleSyncCheck(): void {
     if (this.syncTimer) {
       clearTimeout(this.syncTimer);
     }
 
+    this.currentInterval = this.calculateAdaptiveInterval();
+
     this.syncTimer = setTimeout(() => {
       this.sync().then(() => {
         this.scheduleSyncCheck();
       });
-    }, this.syncInterval);
+    }, this.currentInterval);
+  }
+
+  private deduplicateOperations(operations: QueuedOperation[]): CrdtOperation[] {
+    const dedupMap = new Map<string, QueuedOperation>();
+
+    for (const op of operations) {
+      const existing = dedupMap.get(op.dedupKey);
+      if (!existing || op.timestamp > existing.timestamp) {
+        dedupMap.set(op.dedupKey, op);
+      }
+    }
+
+    return Array.from(dedupMap.values())
+      .sort((a, b) => {
+        if (a.priority !== b.priority) {
+          return a.priority - b.priority;
+        }
+        return a.timestamp - b.timestamp;
+      })
+      .map((qo) => qo.operation);
   }
 
   async sync(): Promise<Result<void, ExtensionError>> {
@@ -133,11 +233,9 @@ export class SyncManager {
     this.isSyncing = true;
 
     try {
-      // Get current operations from queue
-      const operations = [...this.operationQueue];
+      const operations = this.deduplicateOperations(this.operationQueue);
       this.operationQueue = [];
 
-      // Create sync request
       const request: SyncRequest = {
         clock: this.lamportClock,
         device_id: this.deviceId,
@@ -146,33 +244,32 @@ export class SyncManager {
       };
 
       debugLog(
-        `Sending sync request - operations: ${operations.length}, clock: ${this.lamportClock.toString()}, sinceClock: ${this.lastSyncClock?.toString()}`,
+        `Sending adaptive sync - operations: ${operations.length}, interval: ${this.currentInterval}ms`,
       );
 
-      // Call sync endpoint
       const result = await this.api.sync(request);
 
       if (!result.success) {
-        // Re-queue operations on failure
-        this.operationQueue = [...operations, ...this.operationQueue];
+        this.consecutiveErrors++;
+        this.operationQueue = operations.map((op) => ({
+          operation: op,
+          priority: this.getOperationPriority(op),
+          timestamp: Date.now(),
+          dedupKey: this.getOperationDedupKey(op),
+        }));
         return err(result.error);
       }
 
+      this.consecutiveErrors = 0;
       const response = result.data;
 
-      // Update clocks
       this.lamportClock = response.clock;
       this.lastSyncClock = response.clock;
 
-      // Apply remote operations
       await this.applyRemoteOperations(response.operations);
-
-      // Persist state
       await this.persistState();
 
-      debugLog(
-        `Sync completed - remoteOperations: ${response.operations.length}, newClock: ${response.clock.toString()}`,
-      );
+      debugLog(`Adaptive sync completed - remote ops: ${response.operations.length}`);
 
       return ok(undefined);
     } finally {
@@ -191,7 +288,6 @@ export class SyncManager {
   }
 
   private async applyOperation(operation: CrdtOperation): Promise<void> {
-    // Use a custom replacer to handle BigInt
     debugLog(
       `Applying remote operation: ${JSON.stringify(operation, (_, v) => (typeof v === 'bigint' ? v.toString() : v))}`,
     );
@@ -199,23 +295,19 @@ export class SyncManager {
     try {
       switch (operation.type) {
         case 'upsert_tab': {
-          // Check if tab exists
           const tabId = parseInt(operation.id);
           const windowId = parseInt(operation.data.window_id);
 
           try {
-            // Try to get the existing tab
             const tabs = await this.browser.tabs.query({ windowId });
             const existingTab = tabs.find((t) => t.id === tabId);
 
             if (existingTab) {
-              // Update existing tab
               await this.browser.tabs.update(tabId, {
                 url: operation.data.url,
                 active: operation.data.active,
               });
 
-              // Move if needed
               if (existingTab.index !== operation.data.index || existingTab.windowId !== windowId) {
                 await this.browser.tabs.move(tabId, {
                   windowId,
@@ -223,7 +315,6 @@ export class SyncManager {
                 });
               }
             } else {
-              // Create new tab
               await this.browser.tabs.create({
                 windowId,
                 url: operation.data.url,
@@ -242,7 +333,6 @@ export class SyncManager {
           try {
             await this.browser.tabs.remove(tabId);
           } catch {
-            // Tab might already be closed
             debugLog(`Tab ${operation.id} already closed or doesn't exist`);
           }
           break;
@@ -299,14 +389,11 @@ export class SyncManager {
         }
 
         case 'set_window_focus': {
-          // Note: Firefox WebExtension API doesn't support programmatically
-          // focusing windows, so we just log this operation
           debugLog(`Window focus operation for window ${operation.id} (not supported in Firefox)`);
           break;
         }
 
         default: {
-          // Type guard to ensure exhaustive switch
           const _exhaustive: never = operation;
           debugError(`Unknown operation type:`, _exhaustive);
         }
@@ -316,7 +403,89 @@ export class SyncManager {
     }
   }
 
-  // Methods to add operations to queue
+  private getOperationPriority(op: CrdtOperation): OperationPriority {
+    switch (op.type) {
+      case 'close_tab':
+      case 'track_window':
+      case 'untrack_window':
+        return OperationPriority.CRITICAL;
+      case 'upsert_tab':
+      case 'move_tab':
+        return OperationPriority.HIGH;
+      case 'set_active':
+      case 'set_window_focus':
+        return OperationPriority.NORMAL;
+      case 'change_url':
+        return OperationPriority.LOW;
+      default:
+        return OperationPriority.NORMAL;
+    }
+  }
+
+  private getOperationDedupKey(op: CrdtOperation): string {
+    switch (op.type) {
+      case 'upsert_tab':
+      case 'close_tab':
+      case 'set_active':
+      case 'move_tab':
+      case 'change_url':
+        return `${op.type}:${op.id}`;
+      case 'track_window':
+      case 'untrack_window':
+      case 'set_window_focus':
+        return `window:${op.id}`;
+      default: {
+        const _exhaustive: never = op;
+        return `unknown:${(_exhaustive as CrdtOperation).id}`;
+      }
+    }
+  }
+
+  private queueOperation(operation: CrdtOperation, priority: OperationPriority): void {
+    this.incrementClock();
+    this.lastActivityTime = Date.now();
+
+    if (this.operationQueue.length >= this.adaptiveConfig.maxQueueSize) {
+      debugError(`Operation queue full (${this.adaptiveConfig.maxQueueSize}), dropping oldest`);
+      this.operationQueue.shift();
+    }
+
+    this.operationQueue.push({
+      operation,
+      priority,
+      timestamp: Date.now(),
+      dedupKey: this.getOperationDedupKey(operation),
+    });
+
+    this.triggerBatchedSync(priority);
+  }
+
+  private triggerBatchedSync(priority: OperationPriority): void {
+    if (this.batchTimer && this.pendingBatchPriority !== null) {
+      if (priority < this.pendingBatchPriority) {
+        clearTimeout(this.batchTimer);
+        this.batchTimer = null;
+      } else {
+        return;
+      }
+    }
+
+    this.pendingBatchPriority = priority;
+    const delay = this.adaptiveConfig.batchDelays[priority];
+
+    this.batchTimer = setTimeout(() => {
+      this.batchTimer = null;
+      this.pendingBatchPriority = null;
+
+      if (this.syncTimer) {
+        clearTimeout(this.syncTimer);
+      }
+
+      this.sync().then(() => {
+        this.scheduleSyncCheck();
+      });
+    }, delay);
+  }
 
   queueTabUpsert(
     tabId: string,
@@ -326,7 +495,6 @@ export class SyncManager {
     active: boolean,
     index: number,
   ): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'upsert_tab',
       id: tabId,
@@ -339,35 +507,29 @@ export class SyncManager {
         updated_at: BigInt(Date.now()),
       },
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.HIGH);
   }
 
   queueTabClose(tabId: string): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'close_tab',
       id: tabId,
       closed_at: BigInt(Date.now()),
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.CRITICAL);
   }
 
   queueTabActive(tabId: string, active: boolean): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'set_active',
       id: tabId,
       active,
       updated_at: BigInt(Date.now()),
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.NORMAL);
   }
 
   queueTabMove(tabId: string, windowId: string, index: number): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'move_tab',
       id: tabId,
@@ -375,12 +537,10 @@ export class SyncManager {
       index,
       updated_at: BigInt(Date.now()),
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.HIGH);
   }
 
   queueTabUrlChange(tabId: string, url: string, title: string | null): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'change_url',
       id: tabId,
@@ -388,61 +548,40 @@ export class SyncManager {
       title,
       updated_at: BigInt(Date.now()),
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.LOW);
   }
 
   queueWindowTrack(windowId: string, tracked: boolean): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'track_window',
       id: windowId,
       tracked,
       updated_at: BigInt(Date.now()),
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.CRITICAL);
   }
 
   queueWindowUntrack(windowId: string): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'untrack_window',
       id: windowId,
       updated_at: BigInt(Date.now()),
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.CRITICAL);
   }
 
   queueWindowFocus(windowId: string, focused: boolean): void {
-    this.incrementClock();
     const operation: CrdtOperation = {
       type: 'set_window_focus',
       id: windowId,
       focused,
       updated_at: BigInt(Date.now()),
     };
-    this.operationQueue.push(operation);
-    this.triggerImmediateSync();
+    this.queueOperation(operation, OperationPriority.NORMAL);
   }
 
   private incrementClock(): void {
     this.lamportClock += 1n;
-  }
-
-  private triggerImmediateSync(): void {
-    // Cancel current timer and sync immediately
-    if (this.syncTimer) {
-      clearTimeout(this.syncTimer);
-    }
-
-    // Sync after a short delay to batch operations
-    setTimeout(() => {
-      this.sync().then(() => {
-        this.scheduleSyncCheck();
-      });
-    }, 100);
   }
 
   async setupTabEventHandler(): Promise<TabEventHandler> {
